feat: dal request traffic checker
